app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
authlib==1.3.2
httpx==0.27.2
itsdangerous==2.2.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
jinja2

